requirements.txt
PyQt6

platformdirs
pyautogui
pytesseract
requests
aiohttp
beautifulsoup4
googletrans
python-dotenv

# Windows-only
pywin32 ; sys_platform == "win32"

# Linux-only (pynput → evdev)
pynput ; sys_platform != "win32"
